+++ src/cpu/testers/traffic_gen/SConscript
@@ -1,6 +1,6 @@
 # -*- mode:python -*-
 
-# Copyright (c) 2012 ARM Limited
+# Copyright (c) 2012, 2017 ARM Limited
 # All rights reserved.
 #
 # The license below extends only to copyright in the software and shall
@@ -44,7 +44,13 @@ Import('*')
 if env['HAVE_PROTOBUF']:
     SimObject('TrafficGen.py')
 
-    Source('generators.cc')
+    Source('base_gen.cc')
+    Source('dram_gen.cc')
+    Source('dram_rot_gen.cc')
+    Source('idle_gen.cc')
+    Source('linear_gen.cc')
+    Source('random_gen.cc')
+    Source('trace_gen.cc')
     Source('traffic_gen.cc')
 
     DebugFlag('TrafficGen')
